==> CDD_training_aids/New_Stuff/Teeth/2_Normalize.Rmd <==
---
title: "Normalize Data"
author: "Wesley Burr"
date: "29/03/2022"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
library(dplyr)
```

## Normalize Data

We will load each of the data files, identify the 'primary' peak for
Bromobenze, then normalize all Areas to be relative areas to the
reference standard.

```{r}
load("/home/wburr/Rushali_Analysis/New_Stuff/Teeth/Teeth1.rda")
Teeth1m$Area<-as.numeric(Teeth1m$Area)
bromo <- Teeth1m %>% 
           group_by(Sample) %>% 
           subset(Name == "Benzene, bromo-") %>%
           filter(Area == max(Area))
split_Teeth <- Teeth1m %>% 
                group_by(Sample) %>%
                group_split()
for(j in 1:length(split_Teeth)) {
  split_Teeth[[j]]$Area <- split_Teeth[[j]]$Area / unlist(bromo[j, "Area"])
}
Teeth1m <- do.call("rbind", split_Teeth)
save(file = "/home/wburr/Rushali_Analysis/New_Stuff/Teeth/Teeth1n.rda", Teeth1m)
```

This worked, now let's just loop the others to save writing:
